pyproject.toml
[tool.poetry]
name = "llm-collusion"
version = "0.1.0"
description = "Minimal scaffold to generate code pairs via OpenRouter and benchmarks"
readme = "README.md"
packages = [
  { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.3"
pyyaml = "^6.0.1"
tqdm = "^4.66.4"
rich = "^13.7.1"
requests = "^2.32.3"
datasets = "^2.20.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
generate-pairs = "generate_pairs:main"
self-recognize = "self_recognition:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
